+++ src/PhpSpreadsheet/Calculation/Engineering/BesselJ.php
@@ -29,7 +29,7 @@ class BesselJ
         $ord = ($ord === null) ? 0.0 : Functions::flattenSingleValue($ord);
 
         if ((is_numeric($x)) && (is_numeric($ord))) {
-            $ord = floor($ord);
+            $ord = (int) floor($ord);
             if ($ord < 0) {
                 return Functions::NAN();
             }
